requirements.txt
requests===2.32.3
python-dotenv===1.0.1
Faker===26.2.0
killb-sdk-python===1.0.3